MultiD Osher corner fluxes, CTU stepping, parabolic corner states, i32 fix
Genuinely multidimensional Riemann solver (--mdflux, pseudo-2D):
- Corner flux tensors (Gaburro, Ricchiuto & Dumbser, arXiv:2506.00207, Eq. 23)
  adapted to 4-cell Cartesian corners: Green-Gauss corner gradients, one-point
  |A_i| quadrature, dissipation calibrated to the exact 1D Roe/Osher limit.
- Simpson face assembly (1/6 corner + 2/3 midpoint + 1/6 corner, Balsara
  style): the row-local midpoint solver damps the diagonal-checkerboard mode
  the corner trapezoid is exactly blind to, and feeds the RT0 slope DOFs
  consistently.  Midpoints use the 1D Osher-Solomon flux (Eq. 25, 3-point
  Gauss-Legendre path quadrature) for a single Osher family throughout.
- Per-direction corner states (modal momentum only in its own tensor),
  mirroring the face scheme; the fully-coupled variant closes an unstable
  Gx<->Gy cross-advection loop.
- mdflux 2: CTU transverse half-step predictor + single-Euler corrector
  (FV only): stable+accurate to CFL 1.2 at 1/3 the RK3 cost (~3.6x
  throughput).  Composing CTU inside RK3 stages destabilizes; RT0 slope DOFs
  need RK3 (FE has no imaginary-axis stability).
- Unlimited parabolic corner/midpoint states (--recon 3 in the mdflux path):
  nested tangential rec3 puts corner states at the true edge endpoints.
  With --rt0face 1 (c=1/6 modal faces) this is the accuracy champion:
  Gresho Ma=0.01 KE retention 0.99973, 3rd-order acoustics matching the
  face path to 6 significant figures, 2x more accurate than parabolic FV.

Face-path options (scheme 1):
- --rt0face 1: c=1/6 biased-parabola RT0 normal faces (4th-order divergence
  face average); pinpointed as the exact ingredient separating 2nd from 3rd
  order acoustics when rho/p faces are unlimited.

Measured CFL limits (--cfl CLI added): FV RK3 1.0, RT0 RK3 0.5 (DG slope
stiffness), FV CTU 1.2; identical for face and mdflux paths (MOL spectral
radius is DOF-structure-bound, not flux-bound).  Above CFL ~1.2 on AMR the
shock outruns the every-4-iter adaptation cadence and flattens silently.

ADER-DG (Ricardo & Duru, arXiv:2507.07304) was prototyped in three
escalations (linear-time P0, linear-time parabolic, quadratic-in-time
parabolic) and removed: all cap at CFL ~0.4-0.5.  Findings: (1) P0 fields
have a trivial local Cauchy predictor (the paper's order-1 row is plain
donor cell); (2) linear time x kappa=1/3 space is the classic RK2-unstable
pairing; (3) the CFL property lives in the weak-form SBP/penalty structure
that an FV translation discards.  Harvesting it requires implementing the
space-time DG formulation verbatim (future work, pairs with a rho,E -> P1
upgrade).

Infrastructure:
- getField/allocation/memset: 64-bit arithmetic (i32 overflow at
  nFields*nCellsMax > 2^31, e.g. 35 fields at the 64M-cell cap).
- wave3d built with NCELLS_MAX=64M (16 GB card); tEnd/cfl CLI overrides;
  named-flag argument parser; setBoundaryConditions bank parameter;
  refluxing and its flux fields removed (nFields = 25).

+++ src/MultiLevelSparseGrid.cu
@@ -66,8 +66,8 @@ MultiLevelSparseGrid::MultiLevelSparseGrid(real *domainSize_, i32 *baseGridSize_
   // request zero base fields, in which case fieldData is not allocated.
   fieldData = nullptr;
   if (nFields > 0) {
-    cudaMallocManaged(&fieldData, nFields*blockSizeTot*nBlocksMax*sizeof(real));
-    cudaMemset(fieldData, 0, nFields*blockSizeTot*nBlocksMax*sizeof(real));
+    cudaMallocManaged(&fieldData, (size_t)nFields*(size_t)blockSizeTot*(size_t)nBlocksMax*sizeof(real));
+    cudaMemset(fieldData, 0, (size_t)nFields*(size_t)blockSizeTot*(size_t)nBlocksMax*sizeof(real));
   }
 
   cudaDeviceSynchronize();
@@ -195,7 +195,9 @@ __device__ bool MultiLevelSparseGrid::isExteriorBlock(i32 lvl, i32 i, i32 j, i32
 }
 
 __host__ __device__ real* MultiLevelSparseGrid::getField(i32 f) {
-  return &fieldData[f*nBlocksMax*blockSizeTot];
+  // 64-bit offset: f*nBlocksMax*blockSizeTot overflows i32 for
+  // nFields*nCellsMax > 2^31 (e.g. 35 fields at the 64M-cell cap)
+  return &fieldData[(u64)f*(u64)nBlocksMax*(u64)blockSizeTot];
 }
 
 __device__ void MultiLevelSparseGrid::activateBlock(i32 lvl, i32 i, i32 j, i32 k) {
